+++ docker/archlinux/docker_build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-docker build -t "pm3-arch:1.0" .