+++ docker/fedora-37/docker_build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-docker build -t "pm3-fedora-37:1.0" .