+++ docker/homebrew/docker_build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-docker build -t "pm3-brew:1.0" .