+++ docker/parrot-core-latest/docker_run.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-docker run --volume=$(pwd)/../..:/home/rrg/proxmark3 -w /home/rrg/proxmark3 -it pm3-parrotsec-core-latest:1.0